+++ api/api/authentication.py
@@ -133,14 +133,14 @@ def generate_token(user_id=None, rbac_policies=None):
                           wait_for_complete=True,
                           logger=logging.getLogger('wazuh')
                           )
-    result = raise_if_exc(pool.submit(asyncio.run, dapi.distribute_function()).result()).values()
-    token_exp, rbac_mode = list(result)
+    result = raise_if_exc(pool.submit(asyncio.run, dapi.distribute_function()).result()).dikt
     timestamp = int(time())
-    rbac_policies['rbac_mode'] = rbac_mode
+    rbac_policies['rbac_mode'] = result['rbac_mode']
     payload = {
         "iss": JWT_ISSUER,
-        "iat": int(timestamp),
-        "exp": int(timestamp + token_exp),
+        "aud": "Wazuh API REST",
+        "nbf": int(timestamp),
+        "exp": int(timestamp + result['auth_token_exp_timeout']),
         "sub": str(user_id),
         "rbac_policies": rbac_policies
     }
@@ -148,21 +148,21 @@ def generate_token(user_id=None, rbac_policies=None):
     return jwt.encode(payload, generate_secret(), algorithm=JWT_ALGORITHM)
 
 
-def check_token(username, token_iat_time):
+def check_token(username, token_nbf_time):
     """Check the validity of a token with the current time and the generation time of the token.
 
     Parameters
     ----------
     username : str
         Unique username
-    token_iat_time : int
+    token_nbf_time : int
         Issued at time of the current token
     Returns
     -------
     Dict with the result
     """
     with TokenManager() as tm:
-        result = tm.is_token_valid(username=username, token_iat_time=int(token_iat_time))
+        result = tm.is_token_valid(username=username, token_nbf_time=int(token_nbf_time))
 
     return {'valid': result}
 
@@ -180,9 +180,9 @@ def decode_token(token):
     Dict payload ot the token
     """
     try:
-        payload = jwt.decode(token, generate_secret(), algorithms=[JWT_ALGORITHM])
+        payload = jwt.decode(token, generate_secret(), algorithms=[JWT_ALGORITHM], audience='Wazuh API REST')
         dapi = DistributedAPI(f=check_token,
-                              f_kwargs={'username': payload['sub'], 'token_iat_time': payload['iat']},
+                              f_kwargs={'username': payload['sub'], 'token_nbf_time': payload['nbf']},
                               request_type='local_master',
                               is_async=False,
                               wait_for_complete=True,
@@ -203,7 +203,7 @@ def decode_token(token):
         current_rbac_mode = result['rbac_mode']
         current_expiration_time = result['auth_token_exp_timeout']
         if payload['rbac_policies']['rbac_mode'] != current_rbac_mode or \
-                (payload['exp'] - payload['iat']) != current_expiration_time:
+                (payload['exp'] - payload['nbf']) != current_expiration_time:
             raise Unauthorized
 
         return payload